#1 - 패키지 구조 정리
